Refactor spack.platforms to have a clean __init__.py
