+++ lib/spack/spack/platforms/__init__.py
@@ -2,8 +2,7 @@
 # Spack Project Developers. See the top-level COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-import llnl.util.lang
-
+from ._functions import by_name, host, platforms
 from ._platform import Platform
 from .cray import Cray
 from .darwin import Darwin
@@ -15,41 +14,8 @@ __all__ = [
     'Cray',
     'Darwin',
     'Linux',
-    'Test'
+    'Test',
+    'platforms',
+    'host',
+    'by_name'
 ]
-
-#: List of all the platform classes known to Spack
-platforms = [Cray, Darwin, Linux, Test]
-
-
-def host():
-    """Detect and return the platform for this machine or None if detection fails."""
-    for platform_cls in sorted(platforms, key=lambda plt: plt.priority):
-        if platform_cls.detect():
-            return platform_cls()
-    return None
-
-
-@llnl.util.lang.memoized
-def cls_by_name(name):
-    """Return a platform class that corresponds to the given name or None
-    if there is no match.
-
-    Args:
-        name (str): name of the platform
-    """
-    for platform_cls in sorted(platforms, key=lambda plt: plt.priority):
-        if name.replace("_", "").lower() == platform_cls.__name__.lower():
-            return platform_cls
-    return None
-
-
-def by_name(name):
-    """Return a platform object that corresponds to the given name or None
-    if there is no match.
-
-    Args:
-        name (str): name of the platform
-    """
-    platform_cls = cls_by_name(name)
-    return platform_cls() if platform_cls else None
